Create chat crud methods

+++ src/main/java/ru/iteco/test/controller/UserController.java
@@ -15,24 +15,19 @@ public class UserController {
 
     private final UserService userService;
 
-    // ----------------------------   Help method ----------------------------
-
     @GetMapping()
     public List<UserDto> getUsers() {
         return userService.findAll();
     }
 
     @GetMapping("/{id}")
-    public UserDto getUser(@PathVariable("id") Integer id) {
+    public UserDto getUser(@PathVariable("id") Long id) {
         return userService.findById(id);
     }
-
-    // ----------------------------   Help method ----------------------------
 
     @PostMapping("/add")
     @ResponseStatus(HttpStatus.CREATED)
     public Long createNewUser(@RequestBody UserDto userDto) {
         return userService.save(userDto);
     }
-
 }
